# openapi.json
{"openapi":"3.1.0","info":{"title":"Incident Response Backend","version":"0.1.0"},"paths":{"/auth/google/login":{"get":{"tags":["auth"],"summary":"Google Login","operationId":"google_login_auth_google_login_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/google/callback":{"get":{"tags":["auth"],"summary":"Google Callback","operationId":"google_callback_auth_google_callback_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/me":{"get":{"tags":["auth"],"summary":"Get Me","operationId":"get_me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/incidents":{"get":{"tags":["incidents"],"summary":"List Incidents","operationId":"list_incidents_incidents_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/IncidentStatus"},{"type":"null"}],"title":"Status"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/incidents/{incident_id}/analyze":{"post":{"tags":["incidents"],"summary":"Analyze Incident","operationId":"analyze_incident_incidents__incident_id__analyze_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AlertType":{"type":"string","enum":["high_latency","high_errors","service_down","timeout","connection_error","error_response"],"title":"AlertType"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IncidentListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/IncidentResponse"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"limit":{"type":"integer","title":"Limit"},"offset":{"type":"integer","title":"Offset"}},"type":"object","required":["items","total","limit","offset"],"title":"IncidentListResponse"},"IncidentResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"service_name":{"type":"string","title":"Service Name"},"alert_type":{"$ref":"#/components/schemas/AlertType"},"severity":{"$ref":"#/components/schemas/Severity"},"status":{"$ref":"#/components/schemas/IncidentStatus"},"detected_at":{"type":"string","format":"date-time","title":"Detected At"},"resolved_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Resolved At"},"diagnosis":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Diagnosis"},"diagnosis_confidence":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Diagnosis Confidence"},"suggested_action":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Suggested Action"},"diagnosis_reasoning":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Diagnosis Reasoning"},"referenced_postmortem_titles":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Referenced Postmortem Titles"},"diagnosed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Diagnosed At"}},"type":"object","required":["id","service_name","alert_type","severity","status","detected_at","resolved_at"],"title":"IncidentResponse"},"IncidentStatus":{"type":"string","enum":["open","auto_recovered","resolved"],"title":"IncidentStatus","description":"Python member names are UPPERCASE; persisted string values are lowercase.\n\nMust match Alembic/Postgres incident_status exactly: open, auto_recovered, resolved."},"Severity":{"type":"string","enum":["low","medium","high","critical"],"title":"Severity"},"UserResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"email":{"type":"string","format":"email","title":"Email"},"name":{"type":"string","title":"Name"},"role":{"$ref":"#/components/schemas/UserRole"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","email","name","role","created_at"],"title":"UserResponse"},"UserRole":{"type":"string","enum":["user","admin"],"title":"UserRole"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}